[nj] refactoring to add property service
