+++ documents/postman.json
@@ -1,0 +1,1 @@
+{"id":"9cd34daa-0631-246c-301c-5e0d04d2981c","name":"Fang","timestamp":1453100617820,"requests":[{"collectionId":"9cd34daa-0631-246c-301c-5e0d04d2981c","id":"793afba4-19d1-8309-caf8-2a9c2edf298b","name":"Create A New Property","description":"","url":"http://localhost:3000/api/properties","method":"POST","headers":"Content-Type: application/json\n","data":"{\"suburb\":\"Epping\",\"postcode\":\"2030\",\"price\":\"123\",\"address\":\"2 Victoria Street\",\"imageCount\":2}","dataMode":"raw","timestamp":0,"version":2,"time":1456485139412}]}
