+++ package.json
@@ -25,16 +25,22 @@
   "dependencies": {
     "alt": "^0.17.8",
     "async": "^1.5.0",
+    "babel-core": "^6.1.19",
+    "babel-preset-es2015": "^6.1.18",
+    "babel-preset-react": "^6.1.18",
+    "babel-register": "^6.3.13",
+    "babelify": "^7.2.0",
     "body-parser": "^1.14.1",
     "bower": "^1.7.7",
+    "browserify": "^12.0.1",
     "colors": "^1.1.2",
     "compression": "^1.6.0",
     "express": "^4.13.3",
     "gulp": "^3.9.1",
-    "gulp-if": "^2.0.0",
     "gulp-autoprefixer": "^3.1.0",
     "gulp-concat": "^2.6.0",
     "gulp-cssmin": "^0.1.7",
+    "gulp-if": "^2.0.0",
     "gulp-less": "^3.0.3",
     "gulp-plumber": "^1.0.1",
     "gulp-sass": "^2.2.0",
@@ -45,6 +51,7 @@
     "mongoose": "^4.2.5",
     "morgan": "^1.6.1",
     "nodemon": "^1.8.1",
+    "path": "^0.12.7",
     "react": "^0.14.2",
     "react-dom": "^0.14.2",
     "react-router": "^1.0.0",
@@ -53,19 +60,11 @@
     "socket.io": "^1.3.7",
     "swig": "^1.4.2",
     "underscore": "^1.8.3",
-    "xml2js": "^0.4.15",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0",
-    "babelify": "^7.2.0",
-    "browserify": "^12.0.1",
     "watchify": "^3.6.0",
-    "babel-core": "^6.1.19",
-    "babel-preset-es2015": "^6.1.18",
-    "babel-preset-react": "^6.1.18",
-    "babel-register": "^6.3.13"
+    "xml2js": "^0.4.15"
   },
-  "devDependencies": {
-
-  },
+  "devDependencies": {},
   "license": "MIT"
 }
